+++ src/main/kotlin/net/xenotime/network/proto/NetworkPacket.kt
@@ -3,11 +3,11 @@ package net.xenotime.network.proto
 import net.xenotime.network.proto.codec.NetworkProtoReader
 import net.xenotime.network.proto.codec.NetworkProtoWriter
 
-abstract class NetworkPacket<R: NetworkProtoReader , W: NetworkProtoWriter>(private val id: Byte) {
+abstract class NetworkPacket<R: NetworkProtoReader , W: NetworkProtoWriter>(private val id: Int) {
     abstract fun encode(writer: W)
     abstract fun decode(reader: R)
 
-    fun id(): Byte {
+    fun id(): Int {
         return this.id
     }
 }
