+++ src/main/kotlin/net/xenotime/network/wormhole/proto/WormholeProto.kt
@@ -3,10 +3,10 @@ package net.xenotime.network.wormhole.proto
 import net.xenotime.network.wormhole.proto.packet.TestWormholePacket
 import kotlin.reflect.KClass
 
-enum class WormholeProto(private val id: Byte, private val reference: KClass<out WormholePacket>) {
+enum class WormholeProto(private val id: Int, private val reference: KClass<out WormholePacket>) {
     Test(0x0, TestWormholePacket::class);
 
-    fun id(): Byte {
+    fun id(): Int {
         return this.id
     }
